split `DependencyPropertyGenerator` into multiple files
